Resolves #10 The classes have now been extracted from the base extension.ts file
